melos升级7.3.0
7开始使用新的flutter workspace, 有相应调整，
主要是根目录不能放模块了，会很麻烦，
然后每个模块要明确表明自己是workspace的一分子，
还有根项目要一个一个指定所有子模块的路径，不能通配符，

+++ script/plugin_create.sh
@@ -2,28 +2,24 @@
 . "$(dirname $0)/env.sh"
 
 name=$1
-# name 传入了包名就在packages创建包，否则创建修改根项目，
+
+# 检查包名是否提供
 if [ -z "$name" ] || [ "$name" = "." ]; then
-    cd "$ROOT"
-    rm -f pubspec.yaml README.md
-    mv .gitignore .gitignore.bak
-    # 删除各平台代码
-    rm -rf windows/ macos/ linux/ ios/ android/ web/
-    flutter create --org "$organization" --project-name "$project_name" --template=plugin . "${@:2}"
-    dart pub add dev:melos
-    cat .gitignore.bak >>.gitignore
-    rm -f .gitignore.bak
-else
-    # 检查并创建目录
-    if [ ! -d "$packages_dir" ]; then
-        echo "Directory $packages_dir does not exist. Creating..."
-        mkdir -p "$packages_dir"
-    fi
-    cd "$packages_dir"
-    # 删除各平台代码
-    rm -rf windows/ macos/ linux/ ios/ android/ web/
-    flutter create --org "$organization" --template=plugin "$name" "${@:2}"
-    cd "$name"
+    echo "Error: Package name is required"
+    exit 1
+fi
+
+# 检查并创建目录
+if [ ! -d "$packages_dir" ]; then
+    echo "Directory $packages_dir does not exist. Creating..."
+    mkdir -p "$packages_dir"
+fi
+cd "$packages_dir"
+# 删除各平台代码
+rm -rf windows/ macos/ linux/ ios/ android/ web/
+flutter create --org "$organization" --template=plugin "$name" "${@:2}"
+cd "$name"
+if [ -f "$ROOT/LICENSE" ]; then
     cp "$ROOT/LICENSE" .
 fi
 echo 'include: package:flutter_lints/flutter.yaml' >analysis_options.yaml
